requirements.txt
pymupdf
tqdm
sentence-transformers
faiss-cpu
requests
